First functional version of TLS support.


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@16913 ab1d8caa-1f67-47f1-9e81-24633a41865c

+++ sipXrls/src/ResourceInstance.cpp
@@ -222,7 +222,7 @@ void ResourceInstance::notifyEventCallback(const UtlString* dialogHandle,
             else
             {
                // Check that we don't have too many dialogs.
-               if ((int)mXmlDialogs.entries() <
+               if (mXmlDialogs.entries() <
                    getResourceListServer()->getMaxDialogsInResInst())
                {
                   mXmlDialogs.insertKeyAndValue(new UtlString(id),
